% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{approximate_distribution_coefficient_estimate_T3}
\alias{approximate_distribution_coefficient_estimate_T3}
\title{approximate_distribution_coefficient_estimate_T3}
\usage{
approximate_distribution_coefficient_estimate_T3(null_distribution_matrix)
}
\description{
This function input null distribution for generate_null_distribution_T3 and
output estimated coefficient. Should be run before calculating TWT
Please see Setp 4 in our mauscript for more information
}
