% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{clustermap}
\alias{clustermap}
\title{Cluster map}
\usage{
clustermap(
  dat,
  cols = grep("^w", names(dat), value = TRUE),
  i = seq_len(50),
  agg_var = NULL,
  agg_fun = mean,
  as_cor = FALSE,
  k_row = 3,
  annot_vars = c("TYPE", "GENE"),
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{dat}{Summary statistics data.}

\item{cols}{Numeric columns to plot in the heatmap.}

\item{i}{Indices of rows to include. Set to \code{NULL} to include all rows, 
but be warned that this can become very computationally expensive.}

\item{agg_var}{Variable to aggregate data by. 
Set to \code{NULL} to skip this step.}

\item{agg_fun}{Function to aggregate \code{cols} with.}

\item{as_cor}{Show the heatmap as a correlation matrix
instead of a feature x sample matrix.}

\item{k_row}{an integer scalar with the desired number of groups by which to
color the dendrogram's branches in the rows (uses \link[dendextend]{color_branches})
If NA then \link[dendextend]{find_k} is used to deduce the optimal number of clusters.}

\item{annot_vars}{Variables in \code{dat} to include as row-wise annotations.}

\item{show_plot}{Print the plot.}

\item{...}{
  Arguments passed on to \code{\link[heatmaply:heatmaply]{heatmaply::heatmaply}}
  \describe{
    \item{\code{x}}{can either be a heatmapr object, or a numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}
  }}
}
\value{
A named list containing 
an interactive heatmaply object and the data used to create it.
}
\description{
Generate an interactive clustered heatmap of summary statistics data.
}
\examples{
dat <- ThreeWayTest::data_matrix_final
cm <- clustermap(dat = dat)
}
