% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_estimate.R
\name{coefficient_estimate}
\alias{coefficient_estimate}
\title{coefficient_estimate}
\usage{
coefficient_estimate(null_distribution)
}
\arguments{
\item{null_distribution}{Generated by generate_null_distribution_T3}
}
\value{
List of coefficient of alpha, beta and d.
}
\description{
This function estimate the coeffcient of T3 with null distribution 
generated by generate_null_distribution_T3.
Please see Setp 2 in our manuscript for more information.
}
\examples{
null_distribution<-ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
ThreeWayTest::coefficient_estimate(null_distribution[,1])
}
