% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expM}
\alias{expM}
\title{expM}
\usage{
expM(X, e)
}
\description{
metaCCA: this is the method proposed by Cichonska, for more information please see cic_metaCCA#
code website: https://github.com/aalto-ics-kepaco
input value description: 
please put in the description
}
