% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_1kg_genotype_correct_sub}
\alias{final_1kg_genotype_correct_sub}
\title{final_1kg_genotype_correct_sub}
\format{
An object of class \code{data.frame} with 196036 rows and 10 columns.
}
\usage{
data("final_1kg_genotype_correct_sub")
}
\description{
Note that the full \emph{final_1kg_genotype_correct.Rdat} is not 
uploaded because of the size. Here we upload one part-version of this data
\emph{final_1kg_genotype_correct_sub.Rdata} and 
a document on how to access the full data. 
(add description) 
\code{
load(file.path("data","final_1kg_genotype_correct_sub.Rdata"))
usethis::use_data(final_1kg_genotype_correct_sub, overwrite = TRUE)
}
}
\keyword{datasets}
