% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_autoregressive.R
\name{gen_autoregressive}
\alias{gen_autoregressive}
\title{gen_autoregressive}
\usage{
gen_autoregressive(m, rho)
}
\arguments{
\item{m}{Dimension of the matrix.}

\item{rho}{Correlation coefficient.}
}
\value{
An autoregressive covariance matrix.
}
\description{
This is the function for generating an autoregressive covariance matrix.
}
\examples{
cov_mat <- gen_autoregressive(m = 6, rho = 0.5)
}
