% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_list}
\alias{gene_list}
\title{gene_list}
\format{
An object of class \code{character} of length 16656.
}
\usage{
data("gene_list")
}
\description{
(add description) 
\code{
load(file.path("data","gene_list.Rdata"))
usethis::use_data(gene_list, overwrite = TRUE)
}
}
\keyword{datasets}
