% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_null_distribution_T3}
\alias{generate_null_distribution_T3}
\title{generate_null_distribution_T3}
\usage{
generate_null_distribution_T3(m, n, correlation_matrix, cutoff_value)
}
\description{
This function generate the null distribution of T_eta, should be run before calculating TWT
output generated null distribution for function approximate_distribution_coefficient_estimate_T3
Please see Setp 2 in our manuscript for more information
m stantds for the dimension, n is the number of sample size
}
