% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get remote data}
\usage{
get_data(
  file,
  repo = "bschilder/ThreeWayTest",
  tag = "latest",
  storage_dir = tools::R_user_dir(package = "ThreeWayTest", which = "cache"),
  overwrite = TRUE,
  .token = gh::gh_token(),
  check = FALSE
)
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{.token}{GitHub authentication token, see \verb{[gh::gh_token()]}}
}
\description{
Download remotely stored data via \link[piggyback]{pb_download}.
}
\keyword{internal}
