% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_genotype.R
\name{get_full_genotype}
\alias{get_full_genotype}
\title{Get full 1000 Genomes Project genotype data}
\source{
\code{
#### Preprocessing steps ####
dat <- get("final_1kg_genotype_correct")
dat <- data.table::data.table(dat)
tmp <- file.path(tempdir(),"final_1kg_genotype_correct.rds")
saveRDS(dat,tmp)
piggyback::pb_upload(file = tmp, 
                     repo = "bschilder/ThreeWayTest",
                     tag = "v0.0.2")
}
}
\usage{
get_full_genotype(
  file = "final_1kg_genotype_correct.rds",
  tag = "latest",
  overwrite = TRUE,
  .token = gh::gh_token()
)
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{.token}{GitHub authentication token, see \verb{[gh::gh_token()]}}
}
\value{
Large \link[data.table]{data.table} of genotype data.
}
\description{
Download the full version of the \link[ThreeWayTest]{selected_genotype} data, 
which contains genotypes for 2,504 individuals from the 
\href{https://www.internationalgenome.org/}{1000 Genomes Project}.
}
\examples{
full_genotype <- get_full_genotype()
}
