% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{metaCCA}
\alias{metaCCA}
\title{metaCCA}
\usage{
metaCCA(est_genetic_cov, est_pheno_cov, est_XY_cov, N)
}
\arguments{
\item{est_genetic_cov}{Estimated genotype correlation matrix.}

\item{est_pheno_cov}{Estimated phenotype correlation matrix.}

\item{est_XY_cov}{Estimated correlation matrix between genotype and phenotype.}

\item{N}{Sample size.}
}
\value{
A numeric value represents the p value of metaCCA.
}
\description{
This is the main function of metaCCA.
}
\examples{
est_genetic_cov <- diag(nrow = 2,ncol = 2)
est_pheno_cov <- diag(nrow = 2,ncol = 2)
est_XY_cov <- matrix(rnorm(rep(0,4),diag(nrow = 2, ncol = 2)),2,2)
N <- 1000
ThreeWayTest::metaCCA(est_genetic_cov,est_pheno_cov,est_XY_cov,N)
}
