% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MGAS}
\alias{MGAS}
\title{MGAS}
\usage{
MGAS(z_vector, est_genetic_cor, est_pheno_cor)
}
\description{
MGAS: this is the method proposed by Van, for more information, please see 
Van_MGAS#
input value description: 
z_vector is the vector of z_scores
est_genetic_cov is the estimated correlation matrix of genetic variable
est_genetic_cov is the estimated correlation matrix of genetic variable
output value description: 
p_value of MGAS
}
