% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmap_tidygraph.R
\name{networkmap_tidygraph}
\alias{networkmap_tidygraph}
\title{Network map: make tidygraph}
\usage{
networkmap_tidygraph(dat, cols, node_vars)
}
\arguments{
\item{dat}{Summary statistics data.}

\item{cols}{Numeric columns to plot in the heatmap.}

\item{node_vars}{Columns within \code{dat} to set as nodes. 
Each node variable will link to the next node in the character vector.
You can repeat column names to create more connections between nodes.}
}
\value{
\link[tidygraph]{tbl_graph}
}
\description{
Subfunction of \link[ThreeWayTest]{networkmap}.
}
\keyword{internal}
