% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{new_method}
\alias{new_method}
\title{new_method}
\usage{
new_method(
  z_mat,
  est_genetic_cor,
  est_pheno_cor,
  cutoff_value,
  coefficient_matrix
)
}
\description{
This is the method of TWT
input value description:
z_mat is the matrix of z_scores with row stands for number of variants and coloumns standsfor the number of phenotypes.
est_genetic_cor stands for the estimated correlation matrix of genetic variants
est_pheno_cor stands for the estimated correlation matrix of phenotypes
cutoff_value is the set Omega in our paper and the coefficient matrix is calculated based on function approximate_distribution_coefficient_estimate_T3
(number of row stands for the number of SNPs)
(number of coloumn  stands for the number of Phenotypes)
output value description:
final p_value and list of final combined 3 pvalues
}
