% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA_result}
\alias{PCA_result}
\title{PCA_result}
\format{
An object of class \code{list} of length 2.
}
\usage{
data("PCA_result")
}
\description{
Principle components for population stratification used as 
covariates in regression model. Here we use PCA_result$principal.coordinates
as the covariates in our regression model.
\code{
load(file.path("data","PCA_result.Rdata"))
usethis::use_data(PCA_result, overwrite = TRUE)
}
}
\keyword{datasets}
