% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_data.R
\name{postprocess_data}
\alias{postprocess_data}
\title{Post-process results data}
\usage{
postprocess_data(
  dat,
  cols = grep("^w", names(dat), value = TRUE),
  agg_var = NULL,
  agg_fun = mean
)
}
\arguments{
\item{dat}{Summary statistics data.}

\item{cols}{Numeric columns to plot in the heatmap.}

\item{agg_var}{Variable to aggregate data by. 
Set to \code{NULL} to skip this step.}

\item{agg_fun}{Function to aggregate \code{cols} with.}
}
\value{
(Aggregated) \link[data.table]{data.table}
}
\description{
Post-process results data.
}
\examples{
dat <- ThreeWayTest::data_matrix_final
dat2 <- postprocess_data(dat = dat) 
}
