% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quadratic_test}
\alias{quadratic_test}
\title{quadratic test}
\usage{
quadratic_test(z_vector, estimated_cov)
}
\description{
This is the function for method used in real data analysis 
and simulation result.
This is a traditional test, for more information see zhu_Shom
input value description: 
z_vector is the vector of z_scores
estimated_cov is the estimated covariance matrix of these z_scores
output value description: 
p_value of quadratic test
}
