% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selected_genotype}
\alias{selected_genotype}
\title{selected_genotype}
\format{
An object of class \code{data.frame} with 9 rows and 2510 columns.
}
\usage{
data("selected_genotype")
}
\description{
2504 samples in 1000 genomes project for 
calculating covariance matrix of genotypes.
CHR is the chromesome number.
SNP is the number of SNP.
BP is the poisition of the SNP.
Counted is the  allele that counts as 1.
ALT is the allele that counts as 0.
Due to R package limit of data, this is a part-version with 9 SNPs of gene
FADS2 to run getting_start.R in the vigneets.
For full Data, Please go https://github.com/bschilder/ThreeWayTest.
\code{
load(file.path("data","selected_genotype.Rdata"))
usethis::use_data(selected_genotype, overwrite = TRUE)
}
}
\keyword{datasets}
