% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpinfo_12}
\alias{snpinfo_12}
\title{snpinfo_12}
\format{
An object of class \code{data.frame} with 12898 rows and 3 columns.
}
\usage{
data("snpinfo_12")
}
\description{
(add description) 
\code{
snpinfo_12 <- read.table(file.path("data","snpinfo_12.txt"), header = FALSE)
usethis::use_data(snpinfo_12, overwrite = TRUE)
}
}
\keyword{datasets}
