% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{true_cov_mat_generation_autoregressive}
\alias{true_cov_mat_generation_autoregressive}
\title{true_cov_mat_generation_autoregressive}
\usage{
true_cov_mat_generation_autoregressive(m, rho)
}
\description{
Function for covariance matrix generation.
}
