% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWT.R
\name{TWT}
\alias{TWT}
\title{TWT}
\usage{
TWT(z_mat, est_genetic_cor, est_pheno_cor, cutoff_value, coefficient_matrix)
}
\arguments{
\item{z_mat}{is the matrix of z_scores with row number of rows stands for
the number of phenotypes and number of columns
 stands for the number of variants.}

\item{est_genetic_cor}{Estimated correlation matrix of genetic variants.}

\item{est_pheno_cor}{Estimated correlation matrix of phenotypes.}

\item{cutoff_value}{Set Omega.}

\item{coefficient_matrix}{Calculated based on function 
\link[ThreeWayTest]{approximate_distribution_coefficient_estimate_T3}.}
}
\value{
p_value_final P_value of TWT.

p_1 P_value of T_1.

p_2 P_value of T_2.

p_3 P_value of T_3.
}
\description{
This is the method of TWT.
}
\examples{
z_mat<-MASS::mvrnorm(5,mu=rep(0,5), Sigma = diag(nrow = 5, ncol = 5))
null_distribution<-ThreeWayTest::generate_null_distribution_T3(m=25,
n=1000,cov_mat=diag(nrow = 25,ncol= 25), cutoff_value=c(0.2,0.4,0.6,0.8,1))
coefficient_matrix<-
    ThreeWayTest::approximate_distribution_coefficient_estimate_T3(
        null_distribution_matrix = null_distribution)
ThreeWayTest::TWT(z_mat=z_mat, est_genetic_cor=diag(nrow = 5, ncol = 5), 
est_pheno_cor=diag(nrow = 5, ncol = 5), cutoff_value=c(0.2,0.4,0.6,0.8,1), 
coefficient_matrix=coefficient_matrix)
}
