% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/approximate_distribution_coefficient_estimate_T3.R
\name{approximate_distribution_coefficient_estimate_T3}
\alias{approximate_distribution_coefficient_estimate_T3}
\title{approximate_distribution_coefficient_estimate_T3}
\usage{
approximate_distribution_coefficient_estimate_T3(null_distribution_matrix)
}
\arguments{
\item{null_distribution_matrix}{Generated by generate_null_distribution_T3}
}
\value{
Coefficient matrix of alpha, beta and d.
}
\description{
This function estimate the coeffcient of T3 with null distribution 
generated by  generate_null_distribution_T3.
Please see Setp 2 in our manuscript for more information.
}
\examples{
null_distribution<-ThreeWayTest::generate_null_distribution_T3(m=6,n=1000,
cov_mat=diag(nrow = 6, ncol = 6), cutoff_value=c(0.2,0.4,0.6,0.8,1))
coefficient_matrix <- 
    ThreeWayTest::approximate_distribution_coefficient_estimate_T3(
    null_distribution_matrix = null_distribution)
}
