% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_matrix_final}
\alias{data_matrix_final}
\title{data_matrix_final}
\format{
An object of class \code{data.frame} with 196036 rows and 13 columns.
}
\usage{
data("data_matrix_final")
}
\description{
matrix of real data.
(expand description) 
\code{
load(file.path("data","data_matrix_final.Rdata"))
usethis::use_data(data_matrix_final, overwrite = TRUE)
}
}
\keyword{datasets}
