% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_autoregressive}
\alias{gen_autoregressive}
\title{gen_autoregressive}
\usage{
gen_autoregressive(m, rho)
}
\arguments{
\item{m}{Dimension of the matrix.}

\item{rho}{Correlation coefficient.}
}
\value{
A autoregressive covariance matrix.
}
\description{
This is the function for generating autoregressive corvaiance matrix.
}
\examples{
ThreeWayTest::gen_autoregressive(6,0.5)
}
